Added the automatic calculation when creating of the total value on reservaExcursao

+++ Models/ExcursaoModel.cs
@@ -23,8 +23,8 @@ namespace ReserveSystem.Models
 		[Column(TypeName = "DATETIME")]
 		public DateTime Data_Fim { get; set;}
 		[Required,DataType(DataType.Currency), Display(Name = "Preço"), DisplayFormat(DataFormatString = "{0:C}")]
-		[Column(TypeName = "FLOAT")]
-		public float Preco {  get; set;}
+		[Column(TypeName = "DECIMAL")]
+		public decimal Preco {  get; set;}
 		[Required, Display(Name = "Staff ID")]
 		[Column(TypeName = "INTEGER")]
 		public int Staff_Id { get; set; }
